Changed the EAN property's datatype from int to long.

+++ 2SemesterProject.Domain/Models/Product.cs
@@ -10,7 +10,7 @@ namespace _2SemesterProject.Domain.Models
     {
         public int ProductID { get; init; }
         public string Name { get; private set; }
-        public int EAN { get; private set;}
+        public long EAN { get; private set;}
         public string Type { get; private set; }
         public decimal PricePerUnit { get; private set; }
         public int NumberInStock { get; private set; }
